backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
